requirements.txt
Flask>=1.1.1
boto3>=1.9.115
flask_cors>=3.0.8
PyJWT >= 1.7.1